backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
groq==0.11.0
langgraph==0.2.45
langchain==0.3.7
langchain-groq==0.2.1
python-dotenv==1.0.1
